+++ app/parsers/__init__.py
@@ -1,0 +1,1 @@
+"""Parser handlers for stat files."""
