+++ tests/valid/statistics_2017.06.11.160000.dupe.dupe.json
@@ -1,0 +1,294 @@
+{
+	"crewscore": 2550,
+	"escapees": 1,
+	"explosions": [],
+	"mapname": "Box Station",
+	"data_revision": "1.2",
+	"deaths": [{
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 186,
+		"death_y": 46,
+		"death_z": 4,
+		"time_of_death": 289.98,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 195,
+		"death_y": 42,
+		"death_z": 4,
+		"time_of_death": 289.98,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 194,
+		"death_y": 40,
+		"death_z": 4,
+		"time_of_death": 290.31,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 197,
+		"death_y": 39,
+		"death_z": 4,
+		"time_of_death": 290.31,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 126,
+		"death_y": 140,
+		"death_z": 3,
+		"time_of_death": 290.64,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 128,
+		"death_y": 138,
+		"death_z": 3,
+		"time_of_death": 290.64,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 139,
+		"death_y": 131,
+		"death_z": 3,
+		"time_of_death": 290.64,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 118,
+		"death_y": 129,
+		"death_z": 3,
+		"time_of_death": 290.97,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 128,
+		"death_y": 118,
+		"death_z": 3,
+		"time_of_death": 290.97,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 126,
+		"death_y": 118,
+		"death_z": 3,
+		"time_of_death": 290.97,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 126,
+		"death_y": 117,
+		"death_z": 3,
+		"time_of_death": 291.3,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 438,
+		"death_y": 251,
+		"death_z": 2,
+		"time_of_death": 291.3,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}, {
+		"key": null,
+		"assigned_role": null,
+		"special_role": null,
+		"realname": "unknown",
+		"mob_typepath": "/mob/living/carbon/human",
+		"death_x": 132,
+		"death_y": 127,
+		"death_z": 2,
+		"time_of_death": 291.3,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 200,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}],
+	"survivors": [{
+		"key": "pseudonym",
+		"assigned_role": "Assistant",
+		"special_role": null,
+		"realname": "Kameron Pycroft",
+		"mob_typepath": "/mob/living/carbon/human",
+		"time_of_death": 0,
+		"damagevalues": {
+			"BRUTE": 0,
+			"FIRE": 0,
+			"TOXIN": 0,
+			"OXY": 0,
+			"CLONE": 0,
+			"BRAIN": 0
+		}
+	}],
+	"uplink_purchases": [],
+	"badass_bundles": [],
+	"population_polls": [],
+	"blood_spilled": 0,
+	"crates_ordered": 0,
+	"artifacts_discovered": 0,
+	"narsie_corpses_fed": 0,
+	"nuked": 0,
+	"borgs_at_roundend": 0,
+	"heads_at_roundend": 0,
+	"cult_runes_written": 0,
+	"cult_runes_nulled": 0,
+	"cult_runes_fumbled": 0,
+	"cult_converted": 0,
+	"cult_tomes_created": 0,
+	"cult_narsie_summoned": 0,
+	"cult_narsie_corpses_fed": 0,
+	"cult_surviving_cultists": 0,
+	"cult_deconverted": 0,
+	"xeno_eggs_laid": 0,
+	"xeno_faces_hugged": 0,
+	"xeno_faces_protected": 0,
+	"blob_wins": 0,
+	"blob_spawned_blob_players": 0,
+	"blob_spores_spawned": 0,
+	"blob_res_generated": 0,
+	"malf_won": 0,
+	"malf_shunted": 0,
+	"malf_modules": [],
+	"revsquad_won": 0,
+	"revsquad_items": [],
+	"round_start_time": "2017-11-06 22:43:37",
+	"round_end_time": "2017-11-06 22:44:28",
+	"mastermode": "sandbox",
+	"tickermode": "sandbox",
+	"mixed_gamemodes": []
+}
